patch applied https://bugs.eclipse.org/bugs/show_bug.cgi?id=264564

+++ plugins/org.eclipse.xtend/src/org/eclipse/internal/xtend/expression/ast/IntegerLiteral.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2005, 2007 committers of openArchitectureWare and others.
+ * Copyright (c) 2005, 2009 committers of openArchitectureWare and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -11,6 +11,7 @@
 
 package org.eclipse.internal.xtend.expression.ast;
 
+import java.math.BigInteger;
 import java.util.Set;
 
 import org.eclipse.xtend.expression.AnalysationIssue;
@@ -20,6 +21,7 @@ import org.eclipse.xtend.typesystem.Type;
 /**
  * @author Sven Efftinge (http://www.efftinge.de)
  * @author Arno Haase
+ * @author Heiko Behrens
  */
 public class IntegerLiteral extends Literal {
 
@@ -29,7 +31,7 @@ public class IntegerLiteral extends Literal {
 
     @Override
     public Object evaluateInternal(final ExecutionContext ctx) {
-        return new Long(getLiteralValue().getValue());
+        return new BigInteger(getLiteralValue().getValue());
     }
 
     public Type analyzeInternal(final ExecutionContext ctx, final Set<AnalysationIssue> issues) {
